+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Support excluding files from property replacement.
+
 ## Version 2.13.0
 
 * Support escaping placeholders with backslash e.g. `\${MY_ENV_VAR}` will no longer be expanded